loadingFiles

+++ src/main/java/com/example/demo/repos/MessagesRepos.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MessagesRepos extends CrudRepository<Message, Integer> {
+public interface MessagesRepos extends CrudRepository<Message, Long> {
     List<Message> findByTag(String tag);
 }
